+++ swap/src/asb/command.rs
@@ -21,7 +21,6 @@ where
     let json = args.json;
     let disable_timestamp = args.disable_timestamp;
     let testnet = args.testnet;
-    let sled = args.sled;
     let config = args.config;
     let command: RawCommand = args.cmd;
 
@@ -30,7 +29,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::Start { resume_only },
@@ -39,7 +37,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::History,
@@ -48,7 +45,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::WithdrawBtc {
@@ -60,7 +56,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::Balance,
@@ -68,7 +63,6 @@ where
         RawCommand::Config => Arguments {
             testnet,
             json,
-            sled,
             disable_timestamp,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
@@ -77,7 +71,6 @@ where
         RawCommand::ExportBitcoinWallet => Arguments {
             testnet,
             json,
-            sled,
             disable_timestamp,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
@@ -90,7 +83,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::Redeem {
@@ -105,7 +97,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::Cancel { swap_id },
@@ -116,7 +107,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::Refund { swap_id },
@@ -127,7 +117,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::Punish { swap_id },
@@ -136,7 +125,6 @@ where
             testnet,
             json,
             disable_timestamp,
-            sled,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
             cmd: Command::SafelyAbort { swap_id },
@@ -196,7 +184,6 @@ pub struct BitcoinAddressNetworkMismatch {
 pub struct Arguments {
     pub testnet: bool,
     pub json: bool,
-    pub sled: bool,
     pub disable_timestamp: bool,
     pub config_path: PathBuf,
     pub env_config: env::Config,
@@ -258,13 +245,6 @@ pub struct RawArguments {
         help = "Disable timestamping of log messages"
     )]
     pub disable_timestamp: bool,
-
-    #[structopt(
-        short,
-        long = "sled",
-        help = "Forces the asb to use the deprecated sled db if it is available"
-    )]
-    pub sled: bool,
 
     #[structopt(
         long = "config",
@@ -386,7 +366,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
@@ -405,7 +384,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
@@ -424,7 +402,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
@@ -447,7 +424,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
@@ -475,7 +451,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
@@ -502,7 +477,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
@@ -529,7 +503,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
@@ -556,7 +529,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
@@ -577,7 +549,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: true,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
@@ -596,7 +567,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: true,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
@@ -615,7 +585,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: true,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
@@ -640,7 +609,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: true,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
@@ -668,7 +636,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: true,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
@@ -696,7 +663,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: true,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
@@ -724,7 +690,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: true,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
@@ -752,7 +717,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: true,
             json: false,
-            sled: false,
             disable_timestamp: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
@@ -773,7 +737,6 @@ mod tests {
         let expected_args = Arguments {
             testnet: false,
             json: false,
-            sled: false,
             disable_timestamp: true,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,