+++ swap/src/bin/asb.rs
@@ -47,7 +47,6 @@ async fn main() -> Result<()> {
         testnet,
         json,
         disable_timestamp,
-        sled,
         config_path,
         env_config,
         cmd,
@@ -92,10 +91,7 @@ async fn main() -> Result<()> {
         ));
     }
 
-    let db_path = config.data.dir.join("database");
-    let sled_path = config.data.dir.join(db_path);
-
-    let db = open_db(sled_path, config.data.dir.join("sqlite"), sled).await?;
+    let db = open_db(config.data.dir.join("sqlite")).await?;
 
     let seed =
         Seed::from_file_or_generate(&config.data.dir).expect("Could not retrieve/initialize seed");